scripts/bindhosts: detect if all hosts download fail = offline or only some = online (#152)
This patch distinguishes between all downloads fail (which means the phone is likely offline) and keeps the old hosts file then.

But it still fails gracefully if some downloads fail from the the downloaded files (which means there is a problem on the server side or URL etc).

This PR is an attempted fix for https://github.com/bindhosts/bindhosts/issues/134